fix styles outside the editor

+++ src/main/java/mchorse/mappet/client/gui/scripts/themes/GuiThemeEditorOverlayPanel.java
@@ -21,26 +21,7 @@ import java.util.List;
 import java.util.function.Consumer;
 
 public class GuiThemeEditorOverlayPanel extends GuiEditorOverlayPanel<GuiThemeEditorOverlayPanel.SyntaxStyleEntry> {
-    public static final String CODE_SAMPLE =
-            "/* Multi-line\n" +
-                    "   comment test */\n" +
-                    "function main(e) {\n" +
-                    "    // This is a single-line comment\n" +
-                    "    const PI = 3.14;\n" +
-                    "    let hex = 0xFF;\n" +
-                    "    var negative = -42;\n" +
-                    "    var str1 = \"double quotes\";\n" +
-                    "    var str2 = 'single quotes';\n" +
-                    "    var isNull = null;\n" +
-                    "    var isTrue = true && false || undefined;\n" +
-                    "    if (e != null && e.subject() != null) {\n" +
-                    "        this.x = Math.max(e.x + 1, 10);\n" +
-                    "        JSON.stringify(e);\n" +
-                    "        return;\n" +
-                    "    }\n" +
-                    "    prototype.call(this);\n" +
-                    "    customFunc(PI);\n" +
-                    "}";
+    public static final String CODE_SAMPLE = "/* Multi-line\n" + "   comment test */\n" + "function main(e) {\n" + "    // This is a single-line comment\n" + "    const PI = 3.14;\n" + "    let hex = 0xFF;\n" + "    var negative = -42;\n" + "    var str1 = \"double quotes\";\n" + "    var str2 = 'single quotes';\n" + "    var isNull = null;\n" + "    var isTrue = true && false || undefined;\n" + "    if (e != null && e.subject() != null) {\n" + "        this.x = Math.max(e.x + 1, 10);\n" + "        JSON.stringify(e);\n" + "        return;\n" + "    }\n" + "    prototype.call(this);\n" + "    customFunc(PI);\n" + "}";
 
     public GuiIconElement open;
 
@@ -63,6 +44,7 @@ public class GuiThemeEditorOverlayPanel extends GuiEditorOverlayPanel<GuiThemeEd
     public GuiColorElement background;
 
     public GuiCodeEditor preview;
+    private SyntaxStyleEntry prevItem = null;
 
     public GuiThemeEditorOverlayPanel(Minecraft mc) {
         super(mc, IKey.lang("mappet.gui.syntax_theme.main"));
@@ -182,9 +164,8 @@ public class GuiThemeEditorOverlayPanel extends GuiEditorOverlayPanel<GuiThemeEd
         }
 
         for (SyntaxStyleEntry entry : list.getList()) {
-            if (entry.file.getName().equals(Mappet.scriptEditorSyntaxStyle.getFile())) {
+            if (entry.file.getName().equals(Mappet.scriptEditorSyntaxStyle.getFileName())) {
                 pickItem(entry, true);
-
                 break;
             }
         }
@@ -232,15 +213,17 @@ public class GuiThemeEditorOverlayPanel extends GuiEditorOverlayPanel<GuiThemeEd
 
     @Override
     protected void removeItem() {
+        if (list.getList().size() <= 1) return;
         list.getCurrentFirst().file.delete();
-
         super.removeItem();
     }
 
     @Override
     protected void pickItem(SyntaxStyleEntry item, boolean select) {
-        item.save();
-        Mappet.scriptEditorSyntaxStyle.set(item.file.getName(), item.style);
+        if (item == null) return;
+
+        if (prevItem != null) prevItem.save();
+        prevItem = item;
 
         preview.getHighlighter().setStyle(item.style);
         preview.resetHighlight();
@@ -272,10 +255,8 @@ public class GuiThemeEditorOverlayPanel extends GuiEditorOverlayPanel<GuiThemeEd
     @Override
     public void onClose() {
         SyntaxStyleEntry item = list.getCurrentFirst();
-
         item.save();
         Mappet.scriptEditorSyntaxStyle.set(item.file.getName(), item.style);
-
         super.onClose();
     }
